+++ packages/lwc-integration/src/components/props/test-title-prop/title-prop.spec.js
@@ -1,4 +1,5 @@
 const assert = require('assert');
+
 describe('Title public prop', () => {
     const URL = 'http://localhost:4567/title-prop';
     let element;
@@ -8,9 +9,8 @@ describe('Title public prop', () => {
     });
 
     it('should have rendered title property propertly', function () {
-        const text = browser.execute(function () {
-            return document.querySelector('x-child').textContent;
-        });
-        assert.deepEqual(text.value, 'Child title');
+        const element = browser.element('x-child');
+        const text = element.getText();
+        assert.deepEqual(text, 'Child title');
     });
 });