+++ packages/lwc-jest-preset/src/stubs/__tests__/auraInstrumentationStub.test.js
@@ -1,0 +1,34 @@
+// use import like consuming components will, and point to generated file in `dist` folder
+// via jest config
+
+import { perfStart } from 'aura-instrumentation';
+
+describe('auraInstrumentationStub.js', () => {
+    afterEach(() => {
+        jest.resetAllMocks();
+    });
+
+    // all functions are the same implementation so just test one
+    describe('perfStart', () => {
+        it('is an inspectable mock', () => {
+            perfStart('foo');
+            expect(perfStart).toBeCalledWith('foo');
+        });
+
+        it('can be called multiple times', () => {
+            perfStart('foo');
+            perfStart('bar');
+            expect(perfStart).toBeCalledWith('bar');
+        });
+
+        it('can have behavior overridden', () => {
+            perfStart.mockImplementation(() => 'mock return');
+            expect(perfStart()).toBe('mock return');
+        });
+
+        it('can override behavior multiple times in same test', () => {
+            perfStart.mockImplementation(() => 'mock return again');
+            expect(perfStart()).toBe('mock return again');
+        });
+    });
+});