+++ packages/lwc-jest-resolver/src/test/__tests__/x-basic.test.js
@@ -1,0 +1,9 @@
+import { createElement } from 'engine';
+import Basic from '../x-basic';
+
+describe('x-basic', () => {
+    it('loads basic component', () => {
+        const element = createElement('x-basic', { is: Basic });
+        expect(element).toBe(element);
+    });
+});