+++ packages/lwc-jest-resolver/src/test/x-basic.js
@@ -1,0 +1,3 @@
+import { Element } from 'engine';
+
+export default class Basic extends Element {}